move interfaces to models
