+++ src/app/shared/models/timer.ts
@@ -1,0 +1,13 @@
+export interface Timer {
+  start: number
+  end: number
+}
+
+export interface ITimer {
+  seconds: number
+  secondsRemaining: number
+  runTimer: boolean
+  hasStarted: boolean
+  hasFinished: boolean
+  displayTime: string
+}
